feat: Enable versioning in B+ tree index
Update the index.py and bplus_tree.py files to enable storing of version
information in the B+ tree index. This enables inform ation about records
to be updated in the index (when the record is updated) and support
finding records based on a previous value at a specified past version.

Co-authored-by: Eric Edwards <eric.edwarsds02@gmail.com>