fix: Fix incorrect writing of updates after first
For some reason, columns written to pages during an update call which
occurs after the first update to a record (i.e. this update builds off
of an existing tail record) were found to be incorrect during testM1.py.

This places a call to get_partial_record, which updates the bufferpool
and appears to resolve this issue in the tester.